[Gold V] Title: 입국심사, Time: 112 ms, Memory: 2412 KB -BaekjoonHub

+++ 백준/Gold/3079. 입국심사/입국심사.cc
@@ -3,18 +3,18 @@
 #include<algorithm>
 
 using namespace std;
-long long N, M;
-vector<long long>num;
+int N, M;
+vector<int>num;
 int main() {
 	cin >> N >> M;
 	num.resize(N);
-	long long max_n = 0;
+	int max_n = 0;
 	for (int i = 0; i < N; i++) {
 		cin >> num[i];
 		max_n = max(num[i], max_n);
 	}
 
-	unsigned long long start = 0, end = max_n*M, mid = 0, sum = 0,ans;
+	unsigned long long start = 0, end = (long long)max_n*M, mid = 0, sum = 0,ans;
 	while (start <= end) {
 		mid = (start + end) / 2;
 		sum = 0;
